requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.4.0
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
frozenlist==1.4.1
gevent==24.2.1
geventhttpclient==2.0.2
greenlet==3.1.0
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
importlib_metadata==8.5.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.1
pillow==10.4.0
protobuf==4.25.4
pycparser==2.22
Pygments==2.18.0
python-dotenv==1.0.1
python-rapidjson==1.20
python-telegram-bot==21.5
pyzmq==26.2.0
rich==13.8.1
setuptools==74.1.2
sh==2.0.7
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.16.0
yarl==1.11.1
zipp==3.20.1
zope.event==5.0
zope.interface==7.0.3